pyproject.toml
[tool.poetry]
name = "dispatch"
version = "0.1.0"
license = "GPLv3"
description = "Evasive Payload Delivery Server"
authors = ["m8sec"]
readme = "README.md"

[tool.poetry.scripts]
dispatch = "dispatch-server:main"

[tool.poetry.dependencies]
python = "^3.10"
PyJWT = ">=2.10.1"
Flask = ">=3.1.1"
MarkupSafe = ">=3.0.2"
requests = ">=2.32.4"
gevent = ">=25.5.1"
pyOpenSSL = ">=25.1.0"
blinker = ">=1.9.0"
certifi = ">=2025.6.15"
cffi = ">=1.17.1"
charset-normalizer = ">=3.4.2"
click = ">=8.2.1"
cryptography = ">=45.0.4"
greenlet = ">=3.2.3"
idna = ">=3.10"
itsdangerous = ">=2.2.0"
jinja2 = ">=3.1.6"
pip = ">=24.3.1"
pycparser = ">=2.22"
setuptools = ">=80.9.0"
typing-extensions = ">=4.14.0"
urllib3 = ">=2.5.0"
werkzeug = ">=3.1.3"
zope-event = ">=5.0"
zope-interface = ">=7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
